fix: resolve trade data, candlestick, dist, and schedule display issues
- Add serde flatten to Trade and Candlestick structs to capture API
  fields like count_fp, yes_price_dollars, and nested price objects
- Fall back to dollar-string fields in table display for trades and
  candlesticks so count, prices, and OHLC data render correctly
- Make market trade accept a positional ticker arg instead of --ticker
- Default candlestick start_ts/end_ts to last 7 days when omitted
- Support categorical markets in dist command (e.g. "Best AI" events)
- Render exchange schedule as a human-readable table instead of raw JSON
- Format candlestick timestamps as YYYY-MM-DD dates

+++ src/commands/portfolio.rs
@@ -149,7 +149,7 @@ pub async fn execute(client: &KalshiClient, cmd: PortfolioCmd, out: &OutputConfi
             let settle_list = settlements.settlements.unwrap_or_default();
 
             let total_position_count = pos_list.len();
-            let total_contracts: i64 = pos_list.iter().map(|p| p.position.unwrap_or(0).abs()).sum();
+            let total_contracts: i64 = pos_list.iter().map(|p| p.position.unwrap_or(0.0).abs() as i64).sum();
 
             let total_settled = settle_list.len();
